errors in eml files and setting defaults in options

+++ src/edu/stanford/nlp/ling/tokensregex/parser/TokenSequenceParserTokenManager.java
@@ -11,7 +11,6 @@ import edu.stanford.nlp.util.Pair;
 import java.io.Reader;
 import java.io.StringReader;
 import java.util.*;
-import java.lang.RuntimeException;
 
 /** Token Manager. */
 class TokenSequenceParserTokenManager implements TokenSequenceParserConstants
@@ -808,7 +807,7 @@ private int jjMoveNfa_0(int startState, int curPos)
       }
       else
       {
-         int hiByte = (int)(curChar >> 8);
+         int hiByte = (curChar >> 8);
          int i1 = hiByte >> 6;
          long l1 = 1L << (hiByte & 077);
          int i2 = (curChar & 0xff) >> 6;
